feat: add ContentWrapper component and refactor header and footer
