+++ src/components/ContentWrapper.module.css
@@ -1,0 +1,5 @@
+@import "tailwindcss";
+
+.contentWrapper {
+  @apply flex flex-col max-w-7xl grow;
+}
