+++ src/components/footer/Footer.tsx
@@ -2,16 +2,17 @@ import styles from "./Footer.module.css";
 import FooterEmail from "./FooterEmail";
 import FooterLogoAndLinks from "./FooterLogoAndLinks";
 import FooterCopyright from "./FooterCopyright";
+import ContentWrapper from "../ContentWrapper";
 
 function Footer() {
   return (
     <footer className={styles.footer}>
       <div className={styles.footerBg}></div>
-      <div className={styles.footerContent}>
+      <ContentWrapper>
         <FooterEmail />
         <FooterLogoAndLinks />
         <FooterCopyright />
-      </div>
+      </ContentWrapper>
     </footer>
   );
 }
